db:populate rake task for seeding with fake data
The only reason we use a custom rake task is because we follow a
tutorial online that did it this way. The db/seeds.rb file itself states
that it should contain code to set up the *default* values of the
database; considering how we are using this task to create fake test
values (and not default values), it makes sense to keep it out of
db/seeds.rb.