card model serializer
We control the generated JSON from a request like `/cards.json` with a
custom serializer. This is made possible by using the
`active_model_serializers` gem that comes with `ember-rails`.

We delete the call to the `embed` method in `application_serializer.rb`
that was auto-generated because of a deprecation warning by Ember. We
follow the instructions in the warning and configure `embed` globally in
`active_model_serializer_mongoid.rb`.